Adapt code following the TODO notes on lib/utils.bash
The build workflow had to be adapted : Quarkus must be used with a JDK 11 since 2.0.0 (https://quarkus.io/blog/quarkus-2-0-0-final-released/#jdk-11-as-minimal-version) but the default Java version is 8 on macos-latest.

The release workflow has been removed : we do not plan to do releases nor to use GoogleCloudPlatform/release-please-action@v2.